backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai
pytest
httpx
slowapi
google-cloud-translate
pytest-cov
sse-starlette
python-jose[cryptography]
passlib[bcrypt]
safety